Sfntly C++ Cleanup
sfntly/c++ defines several confusing typedefs for standard types.  I
used sed to change them back to the standard types.  This will make it
easer for someone new to this codebase (i.e. everbody) to hack on it.

Also, the CMakeLists.txt file assumes you use either GCC or MSVC.  I
use Clang, so I modified it to give Clang the same options as GCC.

Also, the unit tests assume that googletest is located in
`cpp/ext/gtest`.  I added this to a .gitignore file so that I can
symlink to `..../googletest/googletest`.