Make glyph 126 in NotoSans-Thin.ttf (frac14) render correctly
The internal code says that the two arguments of a composite glyph
should be interpreted as uint16, which positions the 1 of the 1/4
fraction 65536 units too far to the right.

This commit makes all simple glyphs of the frac14 visible. The 4 is
in the wrong position, which will be fixed in a follow-up commit.

Fixes part of #10.